requirements.txt
Flask>=2.3.3
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Werkzeug>=2.3.7
gunicorn>=21.2.0
